requirements.txt
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
scikit-learn>=1.2.2
numpy>=1.23.0
requests>=2.28.0
python-dotenv>=1.0.0
fastapi>=0.95.0
uvicorn>=0.22.0
